+++ simgraph.h
@@ -157,8 +157,8 @@ void display_ddd_box_clip(KOORD_VAL x1, KOORD_VAL y1, KOORD_VAL w, KOORD_VAL h, 
 
 
 // unicode save moving in strings
-int get_next_char(const char* text, int pos);
-int get_prev_char(const char* text, int pos);
+size_t get_next_char(const char* text, size_t pos);
+long get_prev_char(const char* text, long pos);
 
 KOORD_VAL display_get_char_width(utf16 c);
 
@@ -166,7 +166,7 @@ KOORD_VAL display_get_char_width(utf16 c);
 #define proportional_string_width(text)          display_calc_proportional_string_len_width(text, 0x7FFF)
 #define proportional_string_len_width(text, len) display_calc_proportional_string_len_width(text, len)
 // length of a string in pixel
-int display_calc_proportional_string_len_width(const char* text, int len);
+int display_calc_proportional_string_len_width(const char* text, size_t len);
 
 /*
  * len parameter added - use -1 for previous behaviour.
